chore(analytics): move interface AnalyticsQueryFilterDecorator under the analytics_engine package
